+++ sdk/core/Azure.Core/api/Azure.Core.net6.0.cs
@@ -240,7 +240,6 @@ namespace Azure
     {
         protected Response() { }
         public abstract string ClientRequestId { get; set; }
-        public override System.BinaryData Content { get { throw null; } }
         public virtual Azure.Core.ResponseHeaders Headers { get { throw null; } }
         protected internal abstract bool ContainsHeader(string name);
         protected internal abstract System.Collections.Generic.IEnumerable<Azure.Core.HttpHeader> EnumerateHeaders();
