+++ sdk/core/System.ServiceModel.Rest/src/ResponseErrorClassifier.cs
@@ -13,6 +13,11 @@ namespace System.ServiceModel.Rest.Core
         /// </summary>
         public virtual bool IsErrorResponse(PipelineMessage message)
         {
+            if (message.Response is null)
+            {
+                throw new InvalidOperationException("IsErrorResponse must be called on a message where the Result is populated.");
+            }
+
             int statusKind = message.Response.Status / 100;
             return statusKind == 4 || statusKind == 5;
         }
